Jpw/logs

+++ types/types.go
@@ -19,6 +19,6 @@ type HashData struct {
 //StatsData is a JSON object to provide data on total number of calls and average response times
 //Keep Average as a flot to maintain precision.  GetAPIStats will return Average * 10e-6
 type StatsData struct {
-	Total   int     `json:"total"`
-	Average float64 `json:"average"`
+	Total   int   `json:"total"`
+	Average int64 `json:"average"`
 }
